Refactor predicates
